requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
opencv-python>=4.8.0
matplotlib>=3.7.0
streamlit>=1.28.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65.0
